Added a flag for json output to the positive command
Also moved StructuredOutput stuff into its own utility
