+++ src/utilities/structured-output/index.ts
@@ -1,0 +1,12 @@
+export {
+  StructuredOutput,
+  ExampleGroupError,
+  ExampleGroupResult,
+  ExampleGroupWarning,
+  EndpointResult,
+  EndpointSummary,
+  ApiSummary,
+  ApiResults,
+  TestConfig,
+} from './structured-output.interface';
+export { default } from './structured-output.factory';
