+++ src/utilities/structured-output/structured-output.interface.ts
